Struct travis::Client

pub struct Client<C> where
    C: Clone + Connect,  { /* fields omitted */ }

Entry point for all travis operations

Instances of Clients may be cloned.

Methods

impl Client<HttpsConnector<HttpConnector>>

fn oss(credential: Option<Credential>, core: &mut Core) -> Result<Self>

Creates an Travis client for open source github repository builds

fn pro(credential: Option<Credential>, core: &mut Core) -> Result<Self>

Creates a Travis client for private github repository builds

impl<C> Client<C> where
C: Clone + Connect,

fn custom<H>(
host: H,
http: HyperClient<C>,
credential: Option<Credential>,
core: &mut Core
) -> Result<Self> where
H: Into<String>,

Creates a Travis client for hosted versions of travis

fn repos(&self) -> Repos<C>

get a list of repos for the a given owner (user or org)

fn env<'a, R>(&self, slug: R) -> Env<C> where
R: Into<Cow<'a, str>>,

get a ref to an env for a given repo slug

fn builds<'a, R>(&self, slug: R) -> Builds<C> where
R: Into<Cow<'a, str>>,

get a ref builds associated with a repo slug

fn jobs(&self, build_id: usize) -> Jobs<C>

get a ref to jobs associated with a build